Use buffered channel, no need for goroutine.
Both versions expect that reader is consumed, otherwise it'll
eventually get "full". Previously that was handled by many
goroutines waiting to write, which could happen in order different
from original.

New version writes to a buffered channel without order issues,
but runs into a problem if buffered channel gets full.

More code can be added to handle that scenario, but it's not
done to keep things simpler. The reader is always consumed
in the case of ivy, so it shouldn't happen in this context.