Use std{in,out,err} variables consistently in main, usage.
Those std{in,out,err} variables are correctly initialized during package
init time, so they're available for use in both main and usage funcs.

This way, all of stderr output goes to the same place.

In practice, this is unlikely to make a difference because those
errors won't happen (since there are no flags being passed), but
it's still helpful for consistency and readability.